+++ packages/deposit-screener/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Unreleased
 
+- add dummy screening delay option to processor CLI
 - subgraph sync checks if `res.data` undefined
 - use custom implementation of EIP712 signer instead of ethers
 - make EIP712 typehash / signature to use new compressed address encoding